backend/tts_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
pyttsx3
gTTS

